# openapi.yaml
openapi: 3.0.1
info:
  title: Travel addict backend
  description: This is the Travel addict API
  version: 1.0.0
externalDocs:
  description: Github
  url: http://github.com/example
servers:
  - url: https://travel-addict-backend-server.herokuapp.com/
  - url: http://localhost:8080/
tags:
  - name: travel-addict
    description: Everything about travel addict

paths:
  /reisen:
    get:
      tags:
        - reise
      summary: Get all reise
      responses:
        "200":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedReise'
    post:
      tags:
        - reise
      summary: add a new reise
      responses:
        "200":
          description: Successfully added a reise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reise'

  /reisen/{uuid}:
    delete:
      tags:
        - reise
      summary: Deletes a reise by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          description: uuid of the entry to delete
      responses:
        "204":
          description: successful operation

  /reisen/{user_id}:
    delete:
      tags:
        - reise
      summary: Deletes a reise by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          description: uuid of the entry to delete
      responses:
        "204":
          description: successful operation
    get:
      tags:
        - reise
      summary: Get all reise by user id
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
          description: user id of the entry to get
      responses:
        "200":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedReise'

  /login:
    post:
      tags:
        - login
      summary: Log in and get cookie for authentication
      responses:
        "200":
          description: Successfull request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'

components:
  schemas:
    Reise:
      type: object
      properties:
        name:
          type: string
        startDatum:
          type: string
          format: date-time
        endDatum:
          type: string
          format: date-time
        land:
          type: string
    SavedReise:
      allOf:
        - $ref: '#/components/schemas/Reise'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid
    Login:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string